x/monitoringc/client/cli/query_monitoring_history.go
package cli

import (
	"context"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/spf13/cast"
	"github.com/spf13/cobra"

	"github.com/tendermint/spn/x/monitoringc/types"
)

func CmdShowMonitoringHistory() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-monitoring-history [launch-id]",
		Short: "Shows the monitoring history for a launch",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			argLaunchID, err := cast.ToUint64E(args[0])
			if err != nil {
				return err
			}

			params := &types.QueryGetMonitoringHistoryRequest{
				LaunchID: argLaunchID,
			}

			res, err := queryClient.MonitoringHistory(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
